Messed around with units, mostly got rid of float output. Began separating SVG DOM creation from printing. Added edge annotations and arrows.

+++ src/graph.h
@@ -45,6 +45,6 @@ struct Graph {
 };
 
 void dump_dot(const Graph &);
-void dump_flow(const Graph &, float timescale);
+void dump_flow(const Graph &, double timescale);
 
 #endif // DEADLOCK_GRAPH_GRAPH_H_
